Moved MapWidget to own file and show municipalities instead of weather forecast
