+++ src/api/config.js
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 export default {
-	API_BASE_URL: 'https://mobility.api.opendatahub.com/v2',
-	ORIGIN: 'webcomp-generic-map'
+	MOBILITY_API_BASE_URL: 'https://mobility.api.opendatahub.com/v2',
+	TOURISM_API_BASE_URL: 'https://tourism.api.opendatahub.com/v1',
+	ORIGIN: 'webcomp-weather-by-municipality',
 };
