+++ src/api/ninjaApi.js
@@ -5,12 +5,32 @@
 import axios from "axios";
 import config from "./config";
 
-export function callGet(path, params) {
-	console.log("call = " + config.API_BASE_URL + path);
+export function callMobilityGet(path, params) {
+	console.log("call = " + config.MOBILITY_API_BASE_URL + path);
 	console.log("call params = ");
 	console.log(params);
 	return axios
-		.get(config.API_BASE_URL + path, {
+		.get(config.MOBILITY_API_BASE_URL + path, {
+			params: params
+		})
+		.then(function(response) {
+			console.log("call response = ");
+			console.log(response.data);
+			console.log(response.config);
+			return response.data;
+		})
+		.catch(function(error) {
+			console.log(error.response);
+			throw error;
+		});
+}
+
+export function callTourismGet(path, params) {
+	console.log("call = " + config.TOURISM_API_BASE_URL + path);
+	console.log("call params = ");
+	console.log(params);
+	return axios
+		.get(config.TOURISM_API_BASE_URL + path, {
 			params: params
 		})
 		.then(function(response) {
@@ -27,7 +47,7 @@ export function callGet(path, params) {
 
 export async function fetchStations(type) {
 	console.log(type)
-	return callGet("/flat/" + (type || '*'), {
+	return callMobilityGet("/flat/" + (type || '*'), {
 			limit: -1,
 			select: "scode,stype,sname,sorigin,scoordinate,smetadata,pcode",
 			where: "scoordinate.neq.null,sactive.eq.true",
@@ -36,6 +56,25 @@ export async function fetchStations(type) {
 		})
 		.then(response => {
 			this.stations = response.data;
+		})
+		.catch(e => {
+			console.log(e)
+			throw e;
+		});
+}
+
+
+export async function fetchMunicipalities(pageNumber, pageSize) {
+	console.log(pageNumber,pageSize)
+	return callTourismGet("/Municipality/", {
+			limit: -1,
+			select: "Plz,Id,Detail,Region,GpsInfo,Gpstype,Latitude,Longitude,Altitude",
+			where: "odhactive.eq.true,active.eq.true",
+			distinct: true,
+			origin: config.ORIGIN
+		})
+		.then(response => {
+			this.municipalities = response;
 		})
 		.catch(e => {
 			console.log(e)
